+++ CotEditor/Sources/EditorTextView.swift
@@ -9,7 +9,7 @@
 //  ---------------------------------------------------------------------------
 //
 //  © 2004-2007 nakamuxu
-//  © 2014-2020 1024jp
+//  © 2014-2021 1024jp
 //
 //  Licensed under the Apache License, Version 2.0 (the "License");
 //  you may not use this file except in compliance with the License.
@@ -822,14 +822,6 @@ final class EditorTextView: NSTextView, Themable, CurrentLineHighlighting, URLDe
             menu.removeItem(fontMenuItem)
         }
         
-        // add "Inspect Character" menu item if single character is selected
-        if (self.string as NSString).substring(with: self.selectedRange).compareCount(with: 1) == .equal {
-            menu.insertItem(withTitle: "Inspect Character".localized,
-                            action: #selector(showSelectionInfo),
-                            keyEquivalent: "",
-                            at: 1)
-        }
-        
         // add "Copy as Rich Text" menu item
         let copyIndex = menu.indexOfItem(withTarget: nil, andAction: #selector(copy(_:)))
         if copyIndex >= 0 {  // -1 == not found
@@ -1171,10 +1163,6 @@ final class EditorTextView: NSTextView, Themable, CurrentLineHighlighting, URLDe
                 //    disable it to make the state same as `replaceQuotesInSelection(_:)`.
                 return !self.selectedRange.isEmpty
             
-            case #selector(showSelectionInfo):
-                return !self.hasMultipleInsertions &&
-                    (self.string as NSString).substring(with: self.selectedRange).compareCount(with: 1) == .equal
-            
             case #selector(toggleComment):
                 if let menuItem = item as? NSMenuItem {
                     let canComment = self.canUncomment(partly: false)
@@ -1201,6 +1189,13 @@ final class EditorTextView: NSTextView, Themable, CurrentLineHighlighting, URLDe
     
     
     // MARK: Public Accessors
+    
+    /// document object representing the text view contents
+    var document: Document?  {
+        
+        return self.window?.windowController?.document as? Document
+    }
+    
     
     /// tab width in number of spaces
     @objc var tabWidth: Int {
@@ -1347,38 +1342,8 @@ final class EditorTextView: NSTextView, Themable, CurrentLineHighlighting, URLDe
     }
     
     
-    /// display character information by popover
-    @IBAction func showSelectionInfo(_ sender: Any?) {
-        
-        var selectedString = (self.string as NSString).substring(with: self.selectedRange)
-        
-        // apply document's line ending
-        if let documentLineEnding = self.document?.lineEnding,
-            documentLineEnding != .lf, selectedString.detectedLineEnding == .lf
-        {
-            selectedString = selectedString.replacingLineEndings(with: documentLineEnding)
-        }
-        
-        guard let characterInfo = try? CharacterInfo(string: selectedString) else { return }
-        
-        let popoverController = CharacterPopoverController.instantiate(for: characterInfo)
-        let positioningRect = self.boundingRect(for: self.selectedRange)?.insetBy(dx: -4, dy: -4) ?? .zero
-        
-        self.scrollRangeToVisible(self.selectedRange)
-        self.showFindIndicator(for: self.selectedRange)
-        popoverController.showPopover(relativeTo: positioningRect, of: self)
-    }
-    
-    
     
     // MARK: Private Methods
-    
-    /// document object representing the text view contents
-    private var document: Document? {
-        
-        return self.window?.windowController?.document as? Document
-    }
-    
     
     /// update coloring settings
     private func applyTheme() {
@@ -1465,7 +1430,7 @@ final class EditorTextView: NSTextView, Themable, CurrentLineHighlighting, URLDe
         let inset = rate * (visibleRect.height - layoutManager.lineHeight)
         
         // halve inset since the input value will be added to both top and bottom
-        let height = max(floor(inset / 2), Self.textContainerInset.height)
+        let height = max((inset / 2).rounded(.down), Self.textContainerInset.height)
         let diff = height - self.textContainerInset.height
         
         guard diff != 0 else { return }