+++ CotEditor/Sources/PrintTextView.swift
@@ -9,7 +9,7 @@
 //  ---------------------------------------------------------------------------
 //
 //  © 2004-2007 nakamuxu
-//  © 2014-2020 1024jp
+//  © 2014-2021 1024jp
 //
 //  Licensed under the Apache License, Version 2.0 (the "License");
 //  you may not use this file except in compliance with the License.
@@ -210,7 +210,7 @@ final class PrintTextView: NSTextView, Themable, URLDetectable {
         // draw line numbers if needed
         if self.printsLineNumber {
             // prepare text attributes for line numbers
-            let fontSize = round(0.9 * (self.font?.pointSize ?? 12))
+            let fontSize = (0.9 * (self.font?.pointSize ?? 12)).rounded()
             let attrs: [NSAttributedString.Key: Any] = [.font: NSFont.lineNumberFont(ofSize: fontSize),
                                                         .foregroundColor: self.textColor ?? .textColor]