+++ CotEditor/Sources/ThemeMetaDataViewController.swift
@@ -8,7 +8,7 @@
 //
 //  ---------------------------------------------------------------------------
 //
-//  © 2017-2020 1024jp
+//  © 2017-2021 1024jp
 //
 //  Licensed under the Apache License, Version 2.0 (the "License");
 //  you may not use this file except in compliance with the License.
@@ -39,7 +39,7 @@ final class ThemeMetaDataViewController: NSViewController {
     @IBAction func jumpToURL(_ sender: Any?) {
         
         guard
-            let theme = self.representedObject as? Metadata,
+            let theme = self.representedObject as? Theme.Metadata,
             let address = theme.distributionURL,
             let url = URL(string: address)
             else { return NSSound.beep() }